+++ base/cvd/cuttlefish/host/libs/config/config_utils.h
@@ -61,4 +61,6 @@ bool HostSupportsQemuCli();
 
 // Whether to use our local QEMU prebuilt
 bool UseQemuPrebuilt();
+
+std::string GetSeccompPolicyDir();
 }
